Slaves now come up successfully with SSH keys.
These converge successfully and repeatably when doing an initial converge and subsequent converges. Groovy scripts still have a tendency to fail if anything changes like new SSH keys or changing the name of a slave, but the solution to clean up any broken state is just to delete the existing object in Jenkins, and a create works fine (vs. a failed update). As I understand it, the translation to Groovy relies on unique ID hashes of the Ruby object, which obviously changes, causing updates to be impossible (it searches for the new hash ID, which doesn't exist and can't be updated yet).