Fix searches by using node.run_state to pass data
Previously, a search result for the master was saved on a slave node, and the nested tags from the master data on the slave were coming back in a search result. Slave data on the master is still set, but using the IP address and name, instead of the whole node attribute tree.